% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global_config.R
\name{add_db_connections_to_global}
\alias{add_db_connections_to_global}
\title{Reads a simple YAML with required DB connections and adds them to the global config}
\usage{
add_db_connections_to_global(db_config)
}
\arguments{
\item{db_config}{Path to YAML file with simple DB entries (new format!)}
}
\description{
Reads a simple YAML with required DB connections and adds them to the global config
}
