% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open_bronnen_links.R
\name{bagviewer_url}
\alias{bagviewer_url}
\title{URL naar de BAG viewer (Kadaster)}
\usage{
bagviewer_url(id)
}
\arguments{
\item{id}{pandid, adresseerbaarobject, of nummeraanduiding uit het BAG (chr)}
}
\description{
URL naar de BAG viewer (Kadaster)
}
