% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/password_encrypt.R
\name{decrypt_config_file}
\alias{decrypt_config_file}
\title{Decrypt a config file}
\usage{
decrypt_config_file(
  file,
  file_out,
  db_config = NULL,
  secret = Sys.getenv("SHINTO_PASS_SECRET"),
  password_names = "dbpassword"
)
}
\description{
Decrypt a config file
}
