% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bag_geo_download.R
\name{download_gemeente_opendata}
\alias{download_gemeente_opendata}
\title{Download BAG and Geo data for a Gemeente}
\usage{
download_gemeente_opendata(
  gemeente,
  out_path = ".",
  re_download = TRUE,
  cbs_jaar,
  kws = FALSE,
  kws_jaar = 2021,
  rds = TRUE,
  feather = FALSE
)
}
\arguments{
\item{gemeente}{E.g. "Eindhoven"}

\item{out_path}{The relative path to write the datasets to.}

\item{rds}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} Always writes RDS}

\item{feather}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} Unused - feather dependency removed}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}. We now use \code{get_bag_from_cache} and \code{get_geo_from_cache}.
Downloads wijk, buurt, gemeente borders, and the BAG for a Gemeente.
For BAG, writes an RDS into \code{out_path} (a spatial version, sf-dataframe).
For geo, writes e.g. "geo_Eindhoven.rds", a list with components: wijken, buurten, grens.
}
