% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand_dataframe.R
\name{expand_dataframe_split}
\alias{expand_dataframe_split}
\title{Expand a dataframe, by splitting a text field}
\usage{
expand_dataframe_split(data, column = "label", sep = ";")
}
\arguments{
\item{data}{A dataframe}

\item{column}{Which column to use as a basis for splitting}

\item{sep}{The separator in the column, separating the fields}
}
\description{
Expand a dataframe, by splitting a text field
}
\examples{
data <- data.frame(group = c("A","B"), addresses = c("one;two;three","four;five"))
expand_dataframe_split(data, "addresses", sep = ";")
}
