% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bag_from_cache.R
\name{get_bag_from_cache}
\alias{get_bag_from_cache}
\title{Get BAG adres_plus for one or more gemeentes, local or server cache}
\usage{
get_bag_from_cache(gemeentes, cache_folder = "cache", force = FALSE)
}
\arguments{
\item{gemeentes}{Vector of gemeente naam to download/read}

\item{cache_folder}{The \strong{local} folder to use for caching. Ignored on rsconnect,
where a fixed path is used ("/data/bag" is a share on rsconnect dev and prod).}

\item{force}{Download a new copy even when we have it in the cache}
}
\description{
Given one or more gemeente's, read the file \verb{bag_<gemeente>_plus.rds}
either from a local cache folder, or from a shared mount on Rstudio Connect.
If the file does not yet exist, it is downloaded and placed in the cache.
Much preferred to the previous use of e.g. 'data_public/.../bag.rds' because
we can reuse the same (up to date!) copy of the BAG table, and we do not have
to redeploy an application if only the BAG has updated.
}
