% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bag_geo_download.R
\name{get_data_polygon}
\alias{get_data_polygon}
\title{Zoek data binnen een polygon}
\usage{
get_data_polygon(
  polygon,
  con,
  table,
  geocolumn,
  min_overlap = 0.9,
  st_function = c("st_intersects", "st_overlaps", "st_contains")
)
}
\arguments{
\item{polygon}{Sf-kolom, projectie 28992}

\item{con}{Database connectie}

\item{table}{Naam van de table, inclusief schema ("latest.adres")}

\item{geocolumn}{Naam van de kolom met geo-info in de tabel}

\item{min_overlap}{Voor polygonen, alleen diegene die minstens met min_overlap overlappen met de zoek polygoon}

\item{st_function}{Een van 3 spatial predicates}
}
\description{
Zoek data binnen een polygon
}
\details{
Voor geavanceerd gebruik.
}
