% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bag_geo_download.R
\name{get_gemeente_geo}
\alias{get_gemeente_geo}
\title{Download Buurt, Wijk, Gemeente borders}
\usage{
get_gemeente_geo(gemeente, ...)
}
\description{
Downloads from the CSB database. Entry 'CBS' must be present in config file.
}
