% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_geo_from_cache.R
\name{get_geo_from_cache}
\alias{get_geo_from_cache}
\title{Get GEO one or more gemeentes, local or server cache}
\usage{
get_geo_from_cache(gemeentes, cache_folder = "cache")
}
\arguments{
\item{gemeentes}{Vector of gemeente naam to download/read}

\item{cache_folder}{The \strong{local} folder to use for caching. Ignored on rsconnect,
where a fixed path is used ("/data/bag" is a share on rsconnect dev and prod).}
}
\description{
Given one or more gemeente's, read the file \verb{geo_<gemeente>.rds}
either from a local cache folder, or from a shared mount on Rstudio Connect.
If the file does not yet exist, it is downloaded and placed in the cache.
Much preferred to the previous use of e.g. 'data_public/.../geo...rds' because
we can reuse the same (up to date!) copy of the GEO tables, and we do not have
to redeploy an application if only these layers have updated.
}
