% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pc6_data.R
\name{get_pc6_data}
\alias{get_pc6_data}
\title{Get postcode-6 data for postcode(s)}
\usage{
get_pc6_data(pc6, con = NULL, ...)
}
\arguments{
\item{pc6}{The postcode-6 code(s), "1234AB" (can be a vector)}

\item{con}{Connection to CBS database. If missing looks for \code{data_cbs} connection in config.yml}
}
\description{
Get postcode-6 data for postcode(s)
}
