% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pc6_polygon.R
\name{get_pc6_polygon}
\alias{get_pc6_polygon}
\title{Get postcode-6 data for a polygon}
\usage{
get_pc6_polygon(geovlak, con = NULL, ...)
}
\arguments{
\item{geovlak}{An sfc polygon (geo column in an sf dataframe)}

\item{con}{Connection to CBS database. If missing looks for \code{data_cbs} connection in config.yml}
}
\description{
Get postcode-6 data for a polygon
}
