% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bag_geo_download.R
\name{get_perceel_geopunt}
\alias{get_perceel_geopunt}
\alias{get_panden_perceel}
\title{Download het perceel voor een geo-punt}
\usage{
get_perceel_geopunt(pnt, con = NULL, ...)

get_panden_perceel(perceel, min_overlap = 0.9, con = NULL, ...)
}
\arguments{
\item{pnt}{sfc, bv. bag$geopunt}

\item{con}{Optioneel, anders "kadaster" (get_perceel_pand) of "BAGdata" (get_panden_perceel)
in het config bestand.}

\item{perceel}{Geovlak van het perceel (uit het Kadaster, typisch gelezen met get_perceel_geopunt)}

\item{min_overlap}{Zorgt ervoor dat panden die slechts het perceel raken (of door meetfouten net
in het perceel liggen) niet worden meegenomen.}
}
\description{
Vind het perceel uit het Kadaster waarbinnen het geo-punt valt.
Resultaat is in 28992 projectie.

Download alle panden uit het BAG op een perceel.
}
