% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open_bronnen_links.R
\name{google_search_url}
\alias{google_search_url}
\title{URL naar Google search}
\usage{
google_search_url(d)
}
\arguments{
\item{d}{Een rij uit het BAG (sf-dataframe)}
}
\description{
URL naar Google search
}
