% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open_bronnen_links.R
\name{google_streetview_url}
\alias{google_streetview_url}
\title{URL naar Google streetview}
\usage{
google_streetview_url(
  d,
  arrangement = c("11", "12"),
  rotation_angle = 0,
  tilt_angle = 0,
  zoom_level = 0,
  pitch = 5
)
}
\arguments{
\item{d}{Een rij uit het BAG (sf-dataframe)}
}
\description{
URL naar Google streetview
}
