% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_management.R
\name{has_config_entry}
\alias{has_config_entry}
\title{#' Check a config for presence of one or more database connections
#' @export
check_config <- function(name, where, file = "conf/config.yml"){}
\usage{
has_config_entry(name, conf, where)
}
\description{
\if{html}{\out{<div class="sourceCode">}}\preformatted{conf <- read_config(file)

for(nm in name)\{
}\if{html}{\out{</div>}}
}
\details{

}
