% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{make_address_field}
\alias{make_address_field}
\title{Define an address field}
\usage{
make_address_field(data, columns = NULL, template = NULL)
}
\arguments{
\item{data}{Either a dataframe (if columns argument is used), or a vector (if template argument is used).}

\item{columns}{List, see Details.}

\item{template}{Character string. See Details.}
}
\description{
Used to define an address field, as a starting point to match the address in the BAG.
}
\details{
This function can be used in two ways. Either provide a dataframe, and \code{columns} to define
which columns to combine into an address field, or provide a text vector and a "template" defining
how the address information is included.

With the columns argument, provide a list like: list(straat = "straatnaam", huisnummer = "huisnr"),
where the allowed fields are 'straat','huisnummer','woonplaats','huisnummer',
'huisnummertoevoeging', 'postcode'. The right-hand value of the list (between "") is the name of the colum
in the provided dataset.

When \code{template} is used, it is used to define how the address is provided. For example,
with data = "Voorbeeldplein 2 Eendhoven", template should be "{straat}{huisnummer}{woonplaats}".
Use curly braces to separate the fields.
}
