% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bag_geo_download.R
\name{proj_4326}
\alias{proj_4326}
\title{Projecteer data naar lat/long (WG84)}
\usage{
proj_4326(data)
}
\description{
Projecteert naar EPS 4326, voor gebruik in leaflet. Werkt ook voor lege spatial columns,
waar `st_transform` moeite mee heeft. Voor polygonen, haalt ook de z-as weg (anders werkt leaflet niet).
}
