% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bag_geo_download.R
\name{rbind_geo}
\alias{rbind_geo}
\title{Row-bind geo objects}
\usage{
rbind_geo(lis)
}
\description{
Combine multiple geo objects into a single one. Geo objects are read with
\code{get_gemeente_geo}.
}
\examples{
\dontrun{
elke tabel heeft kolom 'gm_naam' voor filtering
geos <- lapply(c("Purmerend","Uitgeest"), get_gemeente_geo)
geo <- rbind_geo(geos)
}
}
