% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adres_lookup.R
\name{search_bag_address}
\alias{search_bag_address}
\title{Search for BAG addresses in a text field}
\usage{
search_bag_address(
  txt,
  bag,
  template = "{straat} {huisnummer}",
  progressbar = TRUE
)
}
\arguments{
\item{txt}{Free text field (vector)}

\item{bag}{BAG dataset (dataframe). In-memory.}

\item{template}{The template to look for in the free text field.}

\item{progressbar}{If TRUE (the default), displays a simple progress bar.}
}
\description{
Searches for addresses in a free text field, which may also contain other text.
For each field, all BAG addresses are searched - so this function is slow for larger queries, and
especially larger BAG datasets.
}
\examples{

\dontrun{
search_bag_address("Er is iets gebeurd op Voorbeeldplein 11",
bag, template = "{straat}{huisnummer}")
}
}
