% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shintobag-package.R
\docType{package}
\name{shintobag-package}
\alias{shintobag}
\alias{shintobag-package}
\title{shintobag: Work with the BAG}
\description{
Match untidy addresses to the BAG, download the BAG, work with the BAG.
}
\keyword{internal}
