% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinto_pin.R
\name{shintoPin}
\alias{shintoPin}
\title{R6 Class with methods for rsconnect data pins}
\description{
Get BAG & CBS data in Shiny applications, fast and without hassle.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{server}}{URL of the Rstudio Connect server}

\item{\code{temp_directory}}{Location to download temporary files}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{shintoPin$new()}}
\item \href{#method-register}{\code{shintoPin$register()}}
\item \href{#method-pin_bag_name}{\code{shintoPin$pin_bag_name()}}
\item \href{#method-pin_cbs_name}{\code{shintoPin$pin_cbs_name()}}
\item \href{#method-pin_bag_cbs_data}{\code{shintoPin$pin_bag_cbs_data()}}
\item \href{#method-get_pin_data}{\code{shintoPin$get_pin_data()}}
\item \href{#method-get_bag_pin}{\code{shintoPin$get_bag_pin()}}
\item \href{#method-get_cbs_pin}{\code{shintoPin$get_cbs_pin()}}
\item \href{#method-clone}{\code{shintoPin$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new shinto rsconnect pin object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{shintoPin$new(server = "https://connect.shintolabs.net", temp_directory = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{server}}{URL of the server (defaults to connect server at shinto Labs)}

\item{\code{temp_directory}}{If NULL, defaults to 'cache' in current working directory}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-register"></a>}}
\if{latex}{\out{\hypertarget{method-register}{}}}
\subsection{Method \code{register()}}{
Register the rsconnect board
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{shintoPin$register(config_file = getOption("path_shinydev_config"))}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{config_file}}{Defaults to `options(path_shinydev_config = "<<filehere>>")`}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-pin_bag_name"></a>}}
\if{latex}{\out{\hypertarget{method-pin_bag_name}{}}}
\subsection{Method \code{pin_bag_name()}}{
Name of the pin for the BAG for a gemeente
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{shintoPin$pin_bag_name(gemeente)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{gemeente}}{E.g. "Eindhoven"}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-pin_cbs_name"></a>}}
\if{latex}{\out{\hypertarget{method-pin_cbs_name}{}}}
\subsection{Method \code{pin_cbs_name()}}{
Name of the pin for the CBS for a gemeente
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{shintoPin$pin_cbs_name(gemeente)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{gemeente}}{E.g. "Eindhoven"}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-pin_bag_cbs_data"></a>}}
\if{latex}{\out{\hypertarget{method-pin_bag_cbs_data}{}}}
\subsection{Method \code{pin_bag_cbs_data()}}{
Create a pin for the BAG and CBS data for a gemeente.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{shintoPin$pin_bag_cbs_data(gemeente)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{gemeente}}{E.g. "Eindhoven"}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_pin_data"></a>}}
\if{latex}{\out{\hypertarget{method-get_pin_data}{}}}
\subsection{Method \code{get_pin_data()}}{
Get a pinned dataset
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{shintoPin$get_pin_data(what)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{what}}{The name of the pin on the rsconnect board}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_bag_pin"></a>}}
\if{latex}{\out{\hypertarget{method-get_bag_pin}{}}}
\subsection{Method \code{get_bag_pin()}}{
Get pinned BAG for a gemeente
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{shintoPin$get_bag_pin(gemeente)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{gemeente}}{E.g. "Eindhoven"}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_cbs_pin"></a>}}
\if{latex}{\out{\hypertarget{method-get_cbs_pin}{}}}
\subsection{Method \code{get_cbs_pin()}}{
Get pinned CBS for a gemeente
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{shintoPin$get_cbs_pin(gemeente)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{gemeente}}{E.g. "Eindhoven"}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{shintoPin$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
