% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_address.R
\name{validate_address}
\alias{validate_address}
\alias{split_adres_field}
\title{Split a merged adres field ("Huisstraat 1A 12") and join with BAG}
\usage{
validate_address(
  data,
  adres_column,
  bag,
  bag_columns = c("adresseerbaarobject_id", "openbareruimtenaam", "huisnummer",
    "huisletter", "huisnummertoevoeging")
)

split_adres_field(x, bag)
}
\arguments{
\item{data}{A dataframe with an adres column}

\item{adres_column}{Name (char) of the adres column in data}

\item{bag}{A BAG extract with columns addresseerbaarobject_id,huisnummer,huisletter,
huisnummertoevoeging}

\item{bag_columns}{Which columns to join onto data}
}
\description{
Takes a dataframe with an adres column, and joins selected columns
from a provided BAG extract (typically for a gemeente).
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
