% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bagSelect_module.R
\name{bagSelectUI}
\alias{bagSelectUI}
\alias{bagSelect}
\title{Bag Select Module}
\usage{
bagSelectUI(id, woonplaats_multiple = FALSE, reset_button = TRUE)

bagSelect(
  input,
  output,
  session,
  bag,
  reset_button = reactive(NULL),
  enkel_adres = TRUE,
  allow_null_woonplaats = TRUE,
  ui_straat = c("autocomplete", "selectInput"),
  ping = reactive(runif(1))
)
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
}
