% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bagSelect_module.R
\name{bagSelectUI}
\alias{bagSelectUI}
\alias{bagSelect}
\title{Bag Select Module}
\usage{
bagSelectUI(id, woonplaats_multiple = FALSE)

bagSelect(input, output, session, bag)
}
\description{
Bag Select Module
}
