% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bag_geo_download.R
\name{download_gemeente_opendata}
\alias{download_gemeente_opendata}
\title{Download BAG and Geo data for a Gemeente}
\usage{
download_gemeente_opendata(
  gemeente,
  out_path = ".",
  re_download = TRUE,
  cbs_jaar = c("2018", "2021"),
  kws = FALSE,
  kws_jaar = 2021
)
}
\arguments{
\item{gemeente}{E.g. "Eindhoven"}

\item{out_path}{The relative path to write the datasets to.}
}
\description{
Downloads wijk, buurt, gemeente borders, and the BAG for a Gemeente.
For BAG, writes an RDS into \code{out_path} (a spatial version, sf-dataframe), and a .feather (
a non-spatial version for very fast reading). For geo, writes e.g. "geo_Eindhoven.rds", a list
with components: wijken, buurten, grens.
}
