% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open_bronnen_links.R
\name{externe_bronnen_lokatie}
\alias{externe_bronnen_lokatie}
\title{Tabel met links naar externe bronnen}
\usage{
externe_bronnen_lokatie(d)
}
\arguments{
\item{d}{Een rij uit het BAG (sf-dataframe)}
}
\description{
Tabel met links naar externe bronnen
}
\details{

}
