% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bag_geo_download.R
\name{get_bag}
\alias{get_bag}
\title{Download the BAG}
\usage{
get_bag(gemeente, con = NULL, ...)
}
\arguments{
\item{gemeente}{Bv. "Rozendaal"}
}
\description{
Download the BAG for one Gemeente at a time. Result is an \code{sf} spatial dataframe.
}
