% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cache_path.R
\name{get_bag_cache_path}
\alias{get_bag_cache_path}
\title{Get path where BAG/GEO cached data is stored}
\usage{
get_bag_cache_path(cache = "cache")
}
\arguments{
\item{cache}{The local cache, normally 'cache'}
}
\description{
Get path where BAG/GEO cached data is stored
}
