% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bag_geo_download.R
\name{get_geo}
\alias{get_geo}
\alias{get_gemeente_geo}
\title{Download Buurt, Wijk, Gemeente grenzen.}
\usage{
get_geo(
  gemeente = NULL,
  what = c("grens", "buurten", "wijken"),
  con = NULL,
  ...
)

get_gemeente_geo(gemeente, ...)
}
\arguments{
\item{gemeente}{Gemeentenaam}

\item{what}{Voor \code{get_geo}, "buurten", "wijken", of "grens"}

\item{con}{Connectie naar de CBS database (als leeg, wordt automatisch aangemaakt)}
}
\description{
Download gemeente, wijk, en buurt grenzen uit de CBS Wijk/Buurt kaart.
}
\details{
De config moet 'CBS' connectie details bevatten.
}
