% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bag_geo_download.R
\name{get_geo}
\alias{get_geo}
\alias{get_gemeente_geo}
\title{Download Buurt, Wijk, Gemeente grenzen.}
\usage{
get_geo(
  gemeente = NULL,
  what = c("grens", "buurten", "wijken"),
  jaar = c("2018", "2021", "2022"),
  kws = FALSE,
  kws_jaar = 2021,
  con = NULL,
  ...
)

get_gemeente_geo(gemeente, jaar, kws = FALSE, kws_jaar = 2021, ...)
}
\arguments{
\item{gemeente}{Gemeentenaam}

\item{what}{Voor \code{get_geo}, "buurten", "wijken", of "grens"}

\item{jaar}{Buurt wijk kaart "2018" of "2021"}

\item{kws}{Kerncijfers toevoegen? (default: FALSE)}

\item{kws_jaar}{Kerncijfers voor welk jaar toevoegen?}

\item{con}{Connectie naar de CBS database (als leeg, wordt automatisch aangemaakt)}
}
\description{
Download gemeente, wijk, en buurt grenzen uit de CBS Wijk/Buurt kaart.
}
\details{
De config moet 'data_cbs' connectie details bevatten (naar de CBS database).
}
