% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gemeente_osm.R
\name{get_gemeente_osm}
\alias{get_gemeente_osm}
\title{Get Gemeente OSM data (amenity, shop)}
\usage{
get_gemeente_osm(gemeente, jaar = "2022", ...)
}
\arguments{
\item{gemeente}{Gemeente}

\item{\dots}{Further args passed to \code{\link[=shinto_db_connection]{shinto_db_connection()}}}
}
\description{
Get Gemeente OSM data (amenity, shop)
}
\details{
Need 'data_osm' config entry in local config file
}
