% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adres_lookup.R
\name{match_bag_address}
\alias{match_bag_address}
\title{Match addresses to the BAG}
\usage{
match_bag_address(x, ...)
}
\arguments{
\item{x}{An address-field object, defined with \code{make_address_field}, see Examples.}

\item{bag}{The BAG for this Gemeente (dataframe).}

\item{bag_columns}{Which columns from the BAG to return as a result.}
}
\description{
Match addresses to the BAG
}
\examples{

\dontrun{

# Example for Nederweert
# 'bag' has to be read in first, using a local copy.

adr <- c("Veldweverstraat 3", "Banendijk 5")

make_address_field(adr, template = "{straat}{huisnummer}") \%>\%
  match_bag_address(., bag = bag)
}

}
