% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bag_geo_download.R
\name{shinto_db_connection}
\alias{shinto_db_connection}
\title{Make a database connection}
\usage{
shinto_db_connection(what, file = "conf/config.yml")
}
\description{
Make a connection with RPostgres to one of the Shinto Labs databases.
}
